abstract Tree on createNode for extends Node data, modified lib/BlueM/BaseTree.php lib/BlueM/Tree.php
