+++ lib/BlueM/Tree.php
@@ -23,11 +23,14 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
 namespace BlueM;
+require_once __DIR__ . '/./lib/BlueM/BaseTree.php';
+require_once __DIR__ . '/./lib/BlueM/Tree/Node.php';
+require_once __DIR__ . '/./lib/BlueM/Tree/InvalidParentException.php';
 
 use BlueM\Tree\InvalidParentException;
 use BlueM\Tree\Node;
+use BlueM\BaseTree;
 
 /**
  * Class for dealing with a tree structure that is constructed by referencing parent IDs
@@ -35,188 +38,8 @@ use BlueM\Tree\Node;
  * @author Carsten Bluem <carsten@bluem.net>
  * @license http://www.opensource.org/licenses/bsd-license.php BSD 2-Clause License
  */
-class Tree
+class Tree extends BaseTree
 {
-    /**
-     * API version
-     *
-     * This number will always be in sync with the first digit of the
-     * release version number.
-     *
-     * @var int
-     */
-    const API = 1;
-
-    /**
-     * @var array
-     */
-    protected $nodes = array();
-
-    /**
-     * @var array
-     */
-    protected $options = array();
-
-    /**
-     * Constructor.
-     *
-     * @param array $data    The data for the tree (array of associative arrays)
-     * @param array $options [optional] Currently, the only supported key is "rootId"
-     *                       (ID of the root node)
-     */
-    public function __construct(array $data, array $options = array())
-    {
-        $this->options = array_change_key_case($options, CASE_LOWER);
-        if (!isset($this->options['rootid'])) {
-            $this->options['rootid'] = 0;
-        }
-
-        $this->build($data);
-    }
-
-    /**
-     * Returns a flat, sorted array of all node objects in the tree.
-     *
-     * @return Node[] Nodes, sorted as if the tree was hierarchical,
-     *                i.e.: the first level 1 item, then the children of
-     *                the first level 1 item (and their children), then
-     *                the second level 1 item and so on.
-     */
-    public function getNodes()
-    {
-        $nodes = array();
-        foreach ($this->nodes[$this->options['rootid']]->getDescendants() as $subnode) {
-            $nodes[] = $subnode;
-        }
-        return $nodes;
-    }
-
-    /**
-     * Returns a single node from the tree, identified by its ID.
-     *
-     * @param int $id Node ID
-     *
-     * @throws \InvalidArgumentException
-     *
-     * @return Node
-     */
-    public function getNodeById($id)
-    {
-        if (empty($this->nodes[$id])) {
-            throw new \InvalidArgumentException("Invalid node primary key $id");
-        }
-        return $this->nodes[$id];
-    }
-
-    /**
-     * Returns an array of all nodes in the root level
-     *
-     * @return Node[] Nodes in the correct order
-     */
-    public function getRootNodes()
-    {
-        return $this->nodes[$this->options['rootid']]->getChildren();
-    }
-
-    /**
-     * Returns the first node for which a specific property's values of all ancestors
-     * and the node are equal to the values in the given argument.
-     *
-     * Example: If nodes have property "name", and on the root level there is a node with
-     * name "A" which has a child with name "B" which has a child which has node "C", you
-     * would get the latter one by invoking getNodeByValuePath('name', ['A', 'B', 'C']).
-     * Comparison is case-sensitive and type-safe.
-     *
-     * @param string $name
-     * @param array  $search
-     *
-     * @return Node|null
-     */
-    public function getNodeByValuePath($name, array $search)
-    {
-        $findNested = function (array $nodes, array $tokens) use ($name, &$findNested) {
-            $token = array_shift($tokens);
-            foreach ($nodes as $node) {
-                $nodeName = $node->get($name);
-                if ($nodeName === $token) {
-                    // Match
-                    if (count($tokens)) {
-                        // Search next level
-                        return $findNested($node->getChildren(), $tokens);
-                    } else {
-                        // We found the node we were looking for
-                        return $node;
-                    }
-                }
-            }
-            return null;
-        };
-
-        return $findNested($this->getRootNodes(), $search);
-    }
-
-    /**
-     * Core method for creating the tree
-     *
-     * @param array $data The data from which to generate the tree
-     *
-     * @throws InvalidParentException
-     */
-    private function build(array $data)
-    {
-        $children = array();
-
-        // Create the root node
-        $this->nodes[$this->options['rootid']] = $this->createNode(
-            array(
-                'id'     => $this->options['rootid'],
-                'parent' => null,
-            )
-        );
-
-        foreach ($data as $row) {
-            $this->nodes[$row['id']] = $this->createNode($row);
-            if (empty($children[$row['parent']])) {
-                $children[$row['parent']] = array($row['id']);
-            } else {
-                $children[$row['parent']][] = $row['id'];
-            }
-        }
-
-        foreach ($children as $pid => $childids) {
-            foreach ($childids as $id) {
-                if ((string)$pid === (string)$id) {
-                    throw new InvalidParentException(
-                        "Node with ID $id references its own ID as parent ID"
-                    );
-                }
-                if (isset($this->nodes[$pid])) {
-                    $this->nodes[$pid]->addChild($this->nodes[$id]);
-                } else {
-                    throw new InvalidParentException(
-                        "Node with ID $id points to non-existent parent with ID $pid"
-                    );
-                }
-            }
-        }
-    }
-
-    /**
-     * Returns a textual representation of the tree
-     *
-     * @return string
-     */
-    public function __toString()
-    {
-        $str = array();
-        foreach ($this->getNodes() as $node) {
-            $indent1st = str_repeat('  ', $node->getLevel() - 1).'- ';
-            $indent    = str_repeat('  ', ($node->getLevel() - 1) + 2);
-            $node      = (string) $node;
-            $str[]     = "$indent1st" . str_replace("\n", "$indent\n  ", $node);
-        }
-        return implode("\n", $str);
-    }
 
     /**
      * Creates and returns a node with the given properties
